Digitise/requirements.txt
streamlit
google-generativeai
python-dotenv
Pillow
pandas
openpyxl
requests
